feat: refactor  my reputation list and update NFT-related components
