+++ src/app/dashboard/achievements/page.js
@@ -14,8 +14,32 @@
  * limitations under the License.
  */
 
-import GainedAchievements from '#/domain/reputation/views/gained-achievements';
+'use client';
+import { ConnectModal } from '@mysten/dapp-kit';
+import { useState } from 'react';
+
+import { NoData } from '@/components/NoData';
+
+import MyReputationList from '#/domain/reputation/views/my-reputation-list';
 
 export default function Page() {
-  return <GainedAchievements />;
+  const [open, setOpen] = useState(false);
+  const [loading, setLoading] = useState(false);
+  const [list, setList] = useState([]);
+
+  return (
+    <div className="pb-12">
+      <ConnectModal
+        open={open}
+        onOpenChange={isOpen => setOpen(isOpen)}
+      />
+      <div className="mb-8 flex items-center justify-between">
+        <h1 className="text-[32px] font-bold">Achievements</h1>
+      </div>
+      <div className="grid grid-cols-2 gap-4">
+        <MyReputationList list={list} setOpen={setOpen} loading={loading} setLoading={setLoading} setList={setList} />
+      </div>
+      {list && list.length === 0 && <NoData />}
+    </div>
+  );
 }
