+++ src/domain/reputation/widgets/nft-skeleton/NftSkeleton.js
@@ -14,4 +14,19 @@
  * limitations under the License.
  */
 
-export { default } from './GainedAchievements';
+import { skeleton } from '@/constants/config';
+
+export default function Skeleton() {
+  return (
+    <div className="flex py-6">
+      <div className={`mr-4 h-[120px] rounded-full ${skeleton} !w-[120px]`} />
+      <div className="bottom-3 flex-1 px-3">
+        <div className={`mb-2 text-lg ${skeleton} h-[28px]`} />
+        <p className="flex items-center text-[13px]">
+          <span className={`mb-1 text-lg ${skeleton} h-[28px]`} />
+        </p>
+        <div className={`mt-5 ${skeleton} h-9 !w-[120px]`} />
+      </div>
+    </div>
+  );
+}
